Modified bookList and addForm components to dispatch the add action from the form
