+++ src/components/AddForm.js
@@ -1,29 +1,48 @@
-import PropTypes from 'prop-types';
+import { useDispatch } from "react-redux";
+import { addBook } from "../redux/books/booksSlice";
+import { useState } from 'react';
+import { v4 as uuidv4 } from 'uuid';
 
-const AddForm = ({
-  handleAdd, handleTitle, handleAuthor, title, author,
-}) => (
+const AddForm = () => {
+  const [title, setTitle] = useState('');
+  const [author, setAuthor] = useState('');
+
+  const dispatch = useDispatch();
+
+  const handleTitleChange = (e) => {
+    setTitle(e.target.value);
+  };
+
+  const handleAuthorChange = (e) => {
+    setAuthor(e.target.value);
+  };
+
+  const handleAdd = (e) => {
+    e.preventDefault();
+    const newBook = {
+      item_id: uuidv4(),
+      title,
+      author,
+    };
+    dispatch(addBook(newBook));
+    setTitle('');
+    setAuthor('');
+  }
+
+  return (
   <form onSubmit={handleAdd}>
     <label htmlFor="bookTitle">
       Title
-      <input type="text" id="bookTitle" placeholder="Title" value={title} onChange={handleTitle} />
+      <input type="text" id="bookTitle" placeholder="Title" value={title} onChange={handleTitleChange} />
     </label>
     <br />
     <label htmlFor="bookAuthor">
       Author
-      <input type="text" id="bookAuthor" placeholder="Author" value={author} onChange={handleAuthor} />
+      <input type="text" id="bookAuthor" placeholder="Author" value={author} onChange={handleAuthorChange} />
     </label>
     <br />
     <button type="submit">Add Book</button>
   </form>
-);
+);}
 
 export default AddForm;
-
-AddForm.propTypes = {
-  handleAdd: PropTypes.func.isRequired,
-  handleTitle: PropTypes.func.isRequired,
-  handleAuthor: PropTypes.func.isRequired,
-  title: PropTypes.string.isRequired,
-  author: PropTypes.string.isRequired,
-};
